[Bronze II] Title: 숫자, Time: 16 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.cc
@@ -10,7 +10,7 @@ int main() {
 		return 0;
 	}
 	if (A > B) {
-		int temp = A;
+		long long temp = A;
 		A = B;
 		B = temp;
 	}
